+++ egui/src/widgets/color_picker.rs
@@ -337,8 +337,10 @@ fn color_picker_hsva_2d(ui: &mut Ui, hsva: &mut Hsva, alpha: Alpha) -> bool {
 pub fn color_edit_button_hsva(ui: &mut Ui, hsva: &mut Hsva, alpha: Alpha) -> Response {
     let pupup_id = ui.auto_id_with("popup");
     let open = ui.memory().is_popup_open(pupup_id);
-    let mut button_response =
-        color_button(ui, (*hsva).into(), open).on_hover_text("Click to edit color");
+    let mut button_response = color_button(ui, (*hsva).into(), open);
+    if ui.style().explanation_tooltips {
+        button_response = button_response.on_hover_text("Click to edit color");
+    }
 
     if button_response.clicked() {
         ui.memory().toggle_popup(pupup_id);