+++ egui/src/widgets/drag_value.rs
@@ -215,14 +215,16 @@ impl<'a> Widget for DragValue<'a> {
                 .min_size(ui.spacing().interact_size); // TODO: find some more generic solution to this
 
             let response = ui.add(button);
-            let response = response
-                .on_hover_cursor(CursorIcon::ResizeHorizontal)
-                .on_hover_text(format!(
+            let mut response = response.on_hover_cursor(CursorIcon::ResizeHorizontal);
+
+            if ui.style().explanation_tooltips {
+                response = response .on_hover_text(format!(
                     "{}{}{}\nDrag to edit or click to enter a value.\nPress 'Shift' while dragging for better control.",
                     prefix,
                     value as f32, // Show full precision value on-hover. TODO: figure out f64 vs f32
                     suffix
                 ));
+            }
 
             if response.clicked() {
                 ui.memory().request_focus(kb_edit_id);